requirements.txt
beautifulsoup4==4.6.3
certifi==2018.8.13
chardet==3.0.4
Django==2.1
django-htmlmin==0.10.0
django-markdownx==2.0.23
gunicorn==19.9.0
html5lib==1.0.1
idna==2.7
Markdown==2.6.11
Pillow==5.2.0
psycopg2==2.7.4
pytz==2018.5
requests==2.19.1
six==1.11.0
urllib3==1.23
webencodings==0.5.1
